Init
This is the version that I made outside of source control. It was also
my first project using React. Maybe I'll actually add a compiling step
to it. I plan on continuing to use this as a testbed for small React
experiments.

Yes, multiple aspects of its design are quite bad. See this as a useful
hack.